상품 조회 개발중 중간 커밋

+++ src/main/java/me/sailer/my_atelier/dto/atelier/AddAtelierRequest.java
@@ -13,7 +13,7 @@ public class AddAtelierRequest {
 
     private String name;
     private String address;
-    private int status;
+    private byte status;
 
     public Atelier toEntity() {
         return Atelier.builder()
