+++ src/main/java/me/sailer/my_atelier/enums/AtelierStatus.java
@@ -1,6 +1,7 @@
 package me.sailer.my_atelier.enums;
 
 import lombok.Getter;
+import me.sailer.my_atelier.enums.common.EnumCommonType;
 
 import java.util.Arrays;
 
@@ -9,21 +10,21 @@ import java.util.Arrays;
  */
 @Getter
 public enum AtelierStatus {
-    NON_EXPOSED("노출안됨", 0),
-    ACTIVE("활성화", 1),
-    TEMP_CLOSED("임시휴업", 2),
-    BANNED("권한박탈", 3),
-    SHUTDOWN("폐업", 4);
+    NON_EXPOSED("노출안됨", (byte)0),
+    ACTIVE("활성화", (byte)1),
+    TEMP_CLOSED("임시휴업", (byte)2),
+    BANNED("권한박탈", (byte)3),
+    SHUTDOWN("폐업", (byte)4);
 
     final private String desc;
-    final private int code;
+    final private byte code;
 
-    AtelierStatus(String desc, int code) {
+    AtelierStatus(String desc, byte code) {
         this.desc = desc;
         this.code = code;
     }
 
-    public static AtelierStatus ofCode(int code) {
+    public static AtelierStatus ofCode(byte code) {
         return Arrays.stream(AtelierStatus.values())
                 .filter(v -> v.getCode() == code)
                 .findAny()
